The GridTest is simply a file I have tested the other three files in. I'm adding it so you know what kinds of errors I'm checking for. We won't need it for the final project.
The GridController will contain one instance of the Grid. It has not been tested.

The Grid class will contain two GridElements. It has not been tested.

The GridElement class has two enums to state what barriers, if any, are in the grid square (left, right, up, down) and if there is anything in a grid square (item, player, etc). It has been tested.